move permission test to separate module
